pyproject.toml
[project]
name = "yukari-discordbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py==2.1.0",
    "python-dotenv>=1.2.1",
]
